requirements.txt
aiogram==3.27.0
aiohttp==3.13.5
aiosqlite==0.22.1
chromadb==1.5.8
pydantic-settings==2.14.0
opendeep==0.1.0
python-dotenv==1.2.2